% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof_tests.R
\name{ad_test}
\alias{ad_test}
\title{Anderson-Darling Test}
\usage{
ad_test(fit)
}
\arguments{
\item{fit}{A distfitr_fit object.}
}
\value{
List with test results.
}
\description{
Test that gives more weight to tail deviations than KS test.
More sensitive to differences in distribution tails.
}
