% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{BIC.distfitr_fit}
\alias{BIC.distfitr_fit}
\title{Extract BIC}
\usage{
\method{BIC}{distfitr_fit}(object, ...)
}
\arguments{
\item{object}{A distfitr_fit object}

\item{...}{Additional arguments (unused)}
}
\value{
Numeric BIC value
}
\description{
Extract BIC
}
\examples{
set.seed(123)
x <- rnorm(100, mean = 5, sd = 2)
fit <- fit_distribution(x, "normal")
BIC(fit)
}
