% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap_bca}
\alias{bootstrap_bca}
\title{BCa (Bias-Corrected and Accelerated) Bootstrap}
\usage{
bootstrap_bca(fit, n_bootstrap, conf_level, parallel, n_cores)
}
\description{
BCa (Bias-Corrected and Accelerated) Bootstrap
}
\keyword{internal}
