% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap_nonparametric}
\alias{bootstrap_nonparametric}
\title{Non-parametric Bootstrap}
\usage{
bootstrap_nonparametric(fit, n_bootstrap, conf_level, parallel, n_cores)
}
\description{
Non-parametric Bootstrap
}
\keyword{internal}
