% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap_parametric}
\alias{bootstrap_parametric}
\title{Parametric Bootstrap}
\usage{
bootstrap_parametric(fit, n_bootstrap, conf_level, parallel, n_cores)
}
\description{
Parametric Bootstrap
}
\keyword{internal}
