% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{calculate_bootstrap_ci}
\alias{calculate_bootstrap_ci}
\title{Calculate Bootstrap Confidence Intervals}
\usage{
calculate_bootstrap_ci(boot_samples, params, conf_level)
}
\description{
Calculate Bootstrap Confidence Intervals
}
\keyword{internal}
