% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_kurtosis}
\alias{calculate_kurtosis}
\title{Calculate Excess Kurtosis}
\usage{
calculate_kurtosis(x)
}
\arguments{
\item{x}{Numeric vector}
}
\description{
Calculate Excess Kurtosis
}
\keyword{internal}
