% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_stats}
\alias{calculate_stats}
\title{Calculate Summary Statistics}
\usage{
calculate_stats(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Named vector of statistics
}
\description{
Calculate Summary Statistics
}
\keyword{internal}
