% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{detect_outliers_likelihood}
\alias{detect_outliers_likelihood}
\title{Likelihood Based Outlier Detection}
\usage{
detect_outliers_likelihood(fit, threshold = NULL)
}
\description{
Likelihood Based Outlier Detection
}
\keyword{internal}
