% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{detect_outliers_mahalanobis}
\alias{detect_outliers_mahalanobis}
\title{Mahalanobis Distance Based Outlier Detection}
\usage{
detect_outliers_mahalanobis(fit, threshold = NULL)
}
\description{
Mahalanobis Distance Based Outlier Detection
}
\keyword{internal}
