% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{detect_outliers_zscore}
\alias{detect_outliers_zscore}
\title{Z-Score Based Outlier Detection}
\usage{
detect_outliers_zscore(fit, threshold = NULL)
}
\description{
Z-Score Based Outlier Detection
}
\keyword{internal}
