% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i18n.R
\name{get_dist_description}
\alias{get_dist_description}
\title{Get Distribution Description (Translated)}
\usage{
get_dist_description(dist_name)
}
\arguments{
\item{dist_name}{Character. Internal distribution name.}
}
\value{
Translated distribution description.
}
\description{
Get the translated description of a distribution.
}
\examples{
get_dist_description("normal")
}
