% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i18n.R
\name{get_dist_name}
\alias{get_dist_name}
\title{Get Distribution Name (Translated)}
\usage{
get_dist_name(dist_name)
}
\arguments{
\item{dist_name}{Character. Internal distribution name.}
}
\value{
Translated distribution name.
}
\description{
Get the translated name of a distribution.
}
\examples{
get_dist_name("normal")
}
