% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i18n.R
\name{get_language}
\alias{get_language}
\title{Get Current Language}
\usage{
get_language()
}
\value{
Character. Current language code.
}
\description{
Get the currently active language.
}
\examples{
get_language()
}
