% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{get_starting_values}
\alias{get_starting_values}
\title{Get Starting Values for Optimization}
\usage{
get_starting_values(data, dist_obj)
}
\description{
Get Starting Values for Optimization
}
\keyword{internal}
