% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{list_distributions}
\alias{list_distributions}
\title{List Available Distributions}
\usage{
list_distributions(type = "all")
}
\arguments{
\item{type}{Character. Filter by type: "all" (default), "continuous",
or "discrete".}
}
\value{
Character vector of distribution names.
}
\description{
Get a list of all available distributions in distfitr.
}
\examples{
# All distributions
list_distributions()

# Only continuous
list_distributions("continuous")
}
