% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i18n.R
\name{load_translations}
\alias{load_translations}
\title{Load Translation File}
\usage{
load_translations(lang = NULL)
}
\arguments{
\item{lang}{Character. Language code.}
}
\value{
List with translations.
}
\description{
Load translations from JSON file for a specific language.
}
\keyword{internal}
