% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i18n.R
\name{locale_format}
\alias{locale_format}
\title{Format with Current Locale}
\usage{
locale_format(x, type = "number", digits = 4)
}
\arguments{
\item{x}{Numeric or character value to format.}

\item{type}{Character. Type of formatting: "number", "percent", "pvalue".}

\item{digits}{Integer. Number of decimal places.}
}
\value{
Formatted string.
}
\description{
Format numbers and text according to current locale.
}
\examples{
locale_format(0.1234, "number", 3)
locale_format(0.05, "pvalue")
}
