% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{logLik.distfitr_fit}
\alias{logLik.distfitr_fit}
\title{Extract Log-Likelihood}
\usage{
\method{logLik}{distfitr_fit}(object, ...)
}
\arguments{
\item{object}{A distfitr_fit object}

\item{...}{Additional arguments (unused)}
}
\value{
An object of class "logLik" containing the log-likelihood value
}
\description{
Extract Log-Likelihood
}
\examples{
set.seed(123)
x <- rnorm(100, mean = 5, sd = 2)
fit <- fit_distribution(x, "normal")
logLik(fit)
}
