% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{nobs.distfitr_fit}
\alias{nobs.distfitr_fit}
\title{Extract Number of Observations}
\usage{
\method{nobs}{distfitr_fit}(object, ...)
}
\arguments{
\item{object}{A distfitr_fit object}

\item{...}{Additional arguments (unused)}
}
\value{
Integer sample size
}
\description{
Extract Number of Observations
}
\examples{
set.seed(123)
x <- rnorm(100, mean = 5, sd = 2)
fit <- fit_distribution(x, "normal")
nobs(fit)
}
