% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{plot.distfitr_diagnostics}
\alias{plot.distfitr_diagnostics}
\title{Plot Diagnostics}
\usage{
\method{plot}{distfitr_diagnostics}(x, ...)
}
\arguments{
\item{x}{A distfitr_diagnostics object}

\item{...}{Additional arguments passed to plot}
}
\description{
Plot Diagnostics
}
