% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{plot_cdf}
\alias{plot_cdf}
\title{Plot Empirical and Fitted CDF}
\usage{
plot_cdf(
  fit,
  main = NULL,
  xlab = "Value",
  ylab = "Cumulative Probability",
  ...
)
}
\description{
Plot Empirical and Fitted CDF
}
\keyword{internal}
