% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{plot_density}
\alias{plot_density}
\title{Plot Histogram with Fitted Density}
\usage{
plot_density(fit, main = NULL, xlab = "Value", ylab = "Density", ...)
}
\description{
Plot Histogram with Fitted Density
}
\keyword{internal}
