% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{plot_pp}
\alias{plot_pp}
\title{Plot P-P Plot}
\usage{
plot_pp(
  fit,
  main = NULL,
  xlab = "Theoretical Probabilities",
  ylab = "Empirical Probabilities",
  ...
)
}
\description{
Plot P-P Plot
}
\keyword{internal}
