% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{plot_qq}
\alias{plot_qq}
\title{Plot Q-Q Plot}
\usage{
plot_qq(
  fit,
  main = NULL,
  xlab = "Theoretical Quantiles",
  ylab = "Sample Quantiles",
  ...
)
}
\description{
Plot Q-Q Plot
}
\keyword{internal}
