% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R, R/fitting_i18n_updates.R
\name{print.distfitr_bootstrap}
\alias{print.distfitr_bootstrap}
\title{Print Bootstrap Results}
\usage{
\method{print}{distfitr_bootstrap}(x, ...)

\method{print}{distfitr_bootstrap}(x, ...)
}
\arguments{
\item{x}{A distfitr_bootstrap object}

\item{...}{Additional arguments (unused)}
}
\description{
Print Bootstrap Results

Print Bootstrap Results with i18n
}
