% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{print.distfitr_distribution}
\alias{print.distfitr_distribution}
\title{Print Distribution Object}
\usage{
\method{print}{distfitr_distribution}(x, ...)
}
\arguments{
\item{x}{A distfitr_distribution object}

\item{...}{Additional arguments (unused)}
}
\description{
Print Distribution Object
}
