% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R, R/fitting_i18n_updates.R
\name{print.distfitr_fit}
\alias{print.distfitr_fit}
\title{Print Fitted Distribution}
\usage{
\method{print}{distfitr_fit}(x, ...)

\method{print}{distfitr_fit}(x, ...)
}
\arguments{
\item{x}{A distfitr_fit object}

\item{...}{Additional arguments (unused)}
}
\description{
Print Fitted Distribution

Print Fitted Distribution with i18n
}
