% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_i18n_updates.R, R/gof_tests.R
\name{print.distfitr_gof}
\alias{print.distfitr_gof}
\title{Print GOF Tests with i18n}
\usage{
\method{print}{distfitr_gof}(x, ...)

\method{print}{distfitr_gof}(x, ...)
}
\arguments{
\item{x}{A distfitr_gof object}

\item{...}{Additional arguments (unused)}
}
\description{
Print GOF Tests with i18n

Print All GOF Tests Results
}
