% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof_tests.R
\name{print.distfitr_gof_test}
\alias{print.distfitr_gof_test}
\title{Print GOF Test Results}
\usage{
\method{print}{distfitr_gof_test}(x, ...)
}
\arguments{
\item{x}{A distfitr_gof_test object}

\item{...}{Additional arguments (unused)}
}
\description{
Print GOF Test Results
}
