% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{residuals.distfitr_fit}
\alias{residuals.distfitr_fit}
\title{Calculate Residuals}
\usage{
\method{residuals}{distfitr_fit}(object, type = "quantile", ...)
}
\arguments{
\item{object}{A distfitr_fit object}

\item{type}{Character. Type of residuals: "quantile" (default), "pearson",
"deviance", or "standardized". Note: Currently only "quantile" is supported.}

\item{...}{Additional arguments (unused)}
}
\value{
Numeric vector of residuals
}
\description{
Calculate Residuals
}
\examples{
set.seed(123)
x <- rnorm(100, mean = 5, sd = 2)
fit <- fit_distribution(x, "normal")
res <- residuals(fit)
head(res)
}
