% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{run_parallel_bootstrap}
\alias{run_parallel_bootstrap}
\title{Run Bootstrap in Parallel}
\usage{
run_parallel_bootstrap(boot_func, n_bootstrap, n_cores)
}
\description{
Run Bootstrap in Parallel
}
\keyword{internal}
