% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{summary.distfitr_fit}
\alias{summary.distfitr_fit}
\title{Summary of Fitted Distribution}
\usage{
\method{summary}{distfitr_fit}(object, ...)
}
\arguments{
\item{object}{A distfitr_fit object}

\item{...}{Additional arguments (unused)}
}
\description{
Summary of Fitted Distribution
}
