% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_data}
\alias{validate_data}
\title{Validate Numeric Data}
\usage{
validate_data(
  x,
  name = "data",
  allow_na = FALSE,
  allow_inf = FALSE,
  min_length = 1
)
}
\arguments{
\item{x}{Numeric vector to validate}

\item{name}{Variable name for error messages}

\item{allow_na}{Logical, allow NA values?}

\item{allow_inf}{Logical, allow infinite values?}

\item{min_length}{Minimum required length}
}
\description{
Validate Numeric Data
}
\keyword{internal}
